+++ project/src/pages/my-list/my-list.tsx
@@ -1,168 +1,133 @@
-import React from 'react';
+
 
 function MyList(): JSX.Element {
   return (
-    <React.Fragment>
-      <div className="visually-hidden">
-        {/*<!-- inject:svg -->*/}
-        <svg xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink">
-          <symbol id="add" viewBox="0 0 19 20">
-            {/*<!-- Generator: Sketch 52.2 (67145) - http://www.bohemiancoding.com/sketch -->*/}
-            <title>+</title>
-            <desc>Created with Sketch.</desc>
-            <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-              <polygon id="+" fill="#EEE5B5" points="10.777832 11.2880859 10.777832 19.5527344 8.41650391 19.5527344 8.41650391 11.2880859 0.627929688 11.2880859 0.627929688 8.92675781 8.41650391 8.92675781 8.41650391 0.662109375 10.777832 0.662109375 10.777832 8.92675781 18.5664062 8.92675781 18.5664062 11.2880859"/>
-            </g>
-          </symbol>
-          <symbol id="full-screen" viewBox="0 0 27 27">
-            <path fillRule="evenodd" clipRule="evenodd" d="M23.8571 0H16V3.14286H23.8571V11H27V3.14286V0H23.8571Z" fill="#FFF9D9" fillOpacity="0.7"/>
-            <path fillRule="evenodd" clipRule="evenodd" d="M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z" fill="#FFF9D9" fillOpacity="0.7"/>
-            <path fillRule="evenodd" clipRule="evenodd" d="M0 3.14286L0 11H3.14286L3.14286 3.14286L11 3.14286V0H3.14286H0L0 3.14286Z" fill="#FFF9D9" fillOpacity="0.7"/>
-            <path fillRule="evenodd" clipRule="evenodd" d="M3.14286 27H11V23.8571H3.14286L3.14286 16H0L0 23.8571V27H3.14286Z" fill="#FFF9D9" fillOpacity="0.7"/>
-          </symbol>
-          <symbol id="in-list" viewBox="0 0 18 14">
-            <path fillRule="evenodd" clipRule="evenodd" d="M2.40513 5.35353L6.1818 8.90902L15.5807 0L18 2.80485L6.18935 14L0 8.17346L2.40513 5.35353Z" fill="#EEE5B5"/>
-          </symbol>
-          <symbol id="pause" viewBox="0 0 14 21">
-            {/*<!-- Generator: Sketch 52.2 (67145) - http://www.bohemiancoding.com/sketch -->*/}
-            <title>Artboard</title>
-            <desc>Created with Sketch.</desc>
-            <g id="Artboard" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-              <polygon id="Line" fill="#EEE5B5" fillRule="nonzero" points="0 -1.11910481e-13 4 -1.11910481e-13 4 21 0 21"/>
-              <polygon id="Line" fill="#EEE5B5" fillRule="nonzero" points="10 -1.11910481e-13 14 -1.11910481e-13 14 21 10 21"/>
-            </g>
-          </symbol>
-        </svg>
-        {/*<!-- endinject -->*/}
-      </div>
+    <div className="user-page">
+      <header className="page-header user-page__head">
+        <div className="logo">
+          <a href="main.html" className="logo__link">
+            <span className="logo__letter logo__letter--1">W</span>
+            <span className="logo__letter logo__letter--2">T</span>
+            <span className="logo__letter logo__letter--3">W</span>
+          </a>
+        </div>
 
-      <div className="user-page">
-        <header className="page-header user-page__head">
-          <div className="logo">
-            <a href="main.html" className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
+        <h1 className="page-title user-page__title">My list <span className="user-page__film-count">9</span></h1>
+        <ul className="user-block">
+          <li className="user-block__item">
+            <div className="user-block__avatar">
+              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
+            </div>
+          </li>
+          <li className="user-block__item">
+            <a href="/" className="user-block__link">Sign out</a>
+          </li>
+        </ul>
+      </header>
 
-          <h1 className="page-title user-page__title">My list <span className="user-page__film-count">9</span></h1>
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a href="/" className="user-block__link">Sign out</a>
-            </li>
-          </ul>
-        </header>
+      <section className="catalog">
+        <h2 className="catalog__title visually-hidden">Catalog</h2>
 
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
+        <div className="catalog__films-list">
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Fantastic Beasts: The Crimes of Grindelwald</a>
+            </h3>
+          </article>
 
-          <div className="catalog__films-list">
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Fantastic Beasts: The Crimes of Grindelwald</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Bohemian Rhapsody</a>
+            </h3>
+          </article>
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Bohemian Rhapsody</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Macbeth</a>
+            </h3>
+          </article>
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Macbeth</a>
-              </h3>
-            </article>
-
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/aviator.jpg" alt="Aviator" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Aviator</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/aviator.jpg" alt="Aviator" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Aviator</a>
+            </h3>
+          </article>
 
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/we-need-to-talk-about-kevin.jpg" alt="We need to talk about Kevin" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">We need to talk about Kevin</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/we-need-to-talk-about-kevin.jpg" alt="We need to talk about Kevin" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">We need to talk about Kevin</a>
+            </h3>
+          </article>
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/what-we-do-in-the-shadows.jpg" alt="What We Do in the Shadows" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">What We Do in the Shadows</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/what-we-do-in-the-shadows.jpg" alt="What We Do in the Shadows" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">What We Do in the Shadows</a>
+            </h3>
+          </article>
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/revenant.jpg" alt="Revenant" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Revenant</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/revenant.jpg" alt="Revenant" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Revenant</a>
+            </h3>
+          </article>
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/johnny-english.jpg" alt="Johnny English" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Johnny English</a>
-              </h3>
-            </article>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/johnny-english.jpg" alt="Johnny English" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Johnny English</a>
+            </h3>
+          </article>
 
 
-            <article className="small-film-card catalog__films-card">
-              <div className="small-film-card__image">
-                <img src="img/shutter-island.jpg" alt="Shutter Island" width="280" height="175" />
-              </div>
-              <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">Shutter Island</a>
-              </h3>
-            </article>
-          </div>
-        </section>
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img src="img/shutter-island.jpg" alt="Shutter Island" width="280" height="175" />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">Shutter Island</a>
+            </h3>
+          </article>
+        </div>
+      </section>
 
-        <footer className="page-footer">
-          <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
+      <footer className="page-footer">
+        <div className="logo">
+          <a href="main.html" className="logo__link logo__link--light">
+            <span className="logo__letter logo__letter--1">W</span>
+            <span className="logo__letter logo__letter--2">T</span>
+            <span className="logo__letter logo__letter--3">W</span>
+          </a>
+        </div>
 
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
-      </div>
-    </React.Fragment>
+        <div className="copyright">
+          <p>© 2019 What to watch Ltd.</p>
+        </div>
+      </footer>
+    </div>
   );
 }
